# Speed up integral images