--- a/mpicbg/src/main/java/mpicbg/ij/integral/DoubleIntegralImage.java
+++ b/mpicbg/src/main/java/mpicbg/ij/integral/DoubleIntegralImage.java
@@ -40,8 +40,10 @@ package mpicbg.ij.integral;
 import ij.process.FloatProcessor;
 import ij.process.ImageProcessor;
 
+
 /**
- * 
+ * A 2d integral image (summed-area table) that allows for fast computation of
+ * the sum of pixel values in a rectangular area.
  *
  * @author Stephan Saalfeld &lt;saalfeld@mpi-cbg.de&gt;
  */
@@ -55,7 +57,14 @@ final public class DoubleIntegralImage implements IntegralImage
 	final private int w1;
 
 	final private double[] sum;
-	
+
+	/**
+	 * Package protected constructor with the original pixel data passed as a parameter.
+	 *
+	 * @param pixels original pixel data to be summed
+	 * @param width width of the original data
+	 * @param height height of the original data
+	 */
 	DoubleIntegralImage( final float[] pixels, final int width, final int height )
 	{
 		this.width = width;
@@ -113,12 +122,22 @@ final public class DoubleIntegralImage implements IntegralImage
 		this.sum = sum;
 	}
 	
-	
+
+	/**
+	 * Construct a new {@link DoubleIntegralImage} from a {@link FloatProcessor}.
+	 *
+	 * @param fp the {@link FloatProcessor} as a source of pixel data
+	 */
 	public DoubleIntegralImage( final FloatProcessor fp )
 	{
 		this( ( float[] )fp.getPixels(), fp.getWidth(), fp.getHeight() );
 	}
-	
+
+	/**
+	 * Construct a new {@link DoubleIntegralImage} from an {@link ImageProcessor}.
+	 *
+	 * @param ip the {@link ImageProcessor} as a source of pixel data
+	 */
 	public DoubleIntegralImage( final ImageProcessor ip )
 	{
 		this.width = ip.getWidth();
@@ -127,37 +146,31 @@ final public class DoubleIntegralImage implements IntegralImage
 		w = width + 1;
 		w1 = w + 1;
 		h = height + 1;
-		
-		final int w2 = w + w;
-		
+
 		final int n = w * height + w;
-		final int n1 = n - w1;
-		final int n2 = n1 - w + 2;
-		
 		sum = new double[ n ];
 
 		/* rows */
-		for ( int i = 0, j = w1; j < n; ++j )
-		{
-			final int end = i + width;
-			double s = sum[ j ] = ip.getf( i );
-			for ( ++i, ++j; i < end; ++i, ++j )
-			{
-				s += ip.getf( i );
-				sum[ j ] = s;
+		for (int j = 1; j < h; ++j) {
+			double rowSum = 0;
+			final int offset = (j - 1) * width;
+			final int offsetSum = j * w + 1;
+
+			for (int i = 0; i < width; ++i) {
+				rowSum += ip.getf(offset + i);
+				sum[offsetSum + i] = rowSum;
 			}
 		}
 		
 		/* columns */
-		for ( int j = w1; j < w2; j -= n1 )
-		{
-			final int end = j + n2;
-			
-			double s = sum[ j ];
-			for ( j += w; j < end; j += w )
-			{
-				s += sum[ j ];
-				sum[ j ] = s;
+		final double[] columnSum = new double[width];
+		for (int j = 1; j < w; ++j) {
+			final int offset = j * w + 1;
+
+			for (int i = 0; i < height; ++i) {
+				final int index = offset + i;
+				columnSum[i] += sum[index];
+				sum[index] = columnSum[i];
 			}
 		}
 	}
@@ -166,12 +179,28 @@ final public class DoubleIntegralImage implements IntegralImage
 	public int getWidth() { return width; }
 	@Override
 	public int getHeight() { return height; }
-	
+
+	/**
+	 * Get the sum of the area from (0, 0) (inclusively) to (x, y) (inclusively).
+	 *
+	 * @param x minimum x coordinate
+	 * @param y minimum y coordinate
+	 * @return the sum of the area in the original pixel data
+	 */
 	public double getDoubleSum( final int x, final int y )
 	{
 		return sum[ y * w + w1 + x ];
 	}
-	
+
+	/**
+	 * Get the sum of the area from (xMin, yMin) (exclusively) to (xMax, yMax) (inclusively).
+	 *
+	 * @param xMin minimum x coordinate
+	 * @param yMin minimum y coordinate
+	 * @param xMax maximum x coordinate
+	 * @param yMax maximum y coordinate
+	 * @return the sum of the area in the original pixel data
+	 */
 	public double getDoubleSum( final int xMin, final int yMin, final int xMax, final int yMax )
 	{
 		final int y1w = yMin * w + w1;