… mpicbg/src/main/java/mpicbg/ij/integral/LongIntegralImage.java
… mpicbg/src/main/java/mpicbg/ij/integral/LongRGBIntegralImage.java
… mpicbg/src/main/java/mpicbg/ij/integral/Scale.java
… mpicbg/src/main/java/mpicbg/ij/integral/Tilt.java