While investigating a performance issue (which turned out to be unrelated to this library) in some downstream code, I found that some minimal changes could provide significant speedup: Integral images are accumulated row-wise and column-wise, where column-wise accumulation took significantly longer than the row-wise one.

This PR changes the memory access pattern of the column-wise accumulation to row-wise access, thereby significantly reducing the run time; see the changes in `IntIntegralImage` as an example. These are the min/max run times for 10 runs of computing the integral image for a 20000x20000 image before and after the changes:

|             | Before         | After          |
| ----------- | -------------- | -------------- |
| Row-wise    | 214ms - 267ms  | 222ms - 269ms  |
| Column-wise | 627ms - 716ms  | 131ms - 141ms  |
| Total       | 841ms - 983ms  | 353ms - 410ms  |

All changes either change the accumulation in exactly the same manner, or are IDE warnings which are automatically fixed. For the former category of changes, I manually checked correctness by using images with elements `i -> 1` and `i -> i + 1`.

Let me know what you think, @axtimwalde