@axtimwalde 2025-05-27
Beautiful! Have you benchmarked small images? 1024^2 or something?